+++ tests/schema_simple.py
@@ -143,7 +143,7 @@ class DataB(dj.Lookup):
 
 
 @schema
-class TestUpdate(dj.Lookup):
+class TTestUpdate(dj.Lookup):
     definition = """
     primary_key     : int
     ---
@@ -154,8 +154,7 @@ class TestUpdate(dj.Lookup):
 
     contents = [
         (0, 'my_string', 0.0, np.random.randn(10, 2)),
-        (1, 'my_other_string', 1.0, np.random.randn(20, 1)),
-    ]
+        (1, 'my_other_string', 1.0, np.random.randn(20, 1))]
 
 
 @schema