+++ tests/test_blob2.py
@@ -1,6 +1,6 @@
 import numpy as np
 import datajoint as dj
-from nose.tools import assert_equal, assert_true, assert_list_equal, assert_tuple_equal
+from nose.tools import assert_equal, assert_true, assert_list_equal, assert_tuple_equal, assert_false
 
 from . import PREFIX, CONN_INFO
 
@@ -49,11 +49,18 @@ def insert_blobs():
 
 
 class TestFetch:
-    def __init__(self):
+
+    @classmethod
+    def setup_class(cls):
+        assert_false(dj.config['safemode'], 'safemode must be disabled')
         Blob().delete()
         insert_blobs()
 
-    def test_complex_matlab_blobs(self):
+    @staticmethod
+    def test_complex_matlab_blobs():
+        """
+        test correct de-serialization of various blob types
+        """
         blobs = Blob().fetch('blob', order_by='id')
         assert_equal(blobs[0][0], 'character string')
         assert_true(np.array_equal(blobs[1][0], np.r_[1:180:15]))
@@ -66,3 +73,13 @@ class TestFetch:
         assert_true(blobs[5].dtype == 'uint8')
         assert_tuple_equal(blobs[6].shape, (2, 3, 4))
         assert_true(blobs[6].dtype == 'complex128')
+
+    @staticmethod
+    def test_iter():
+        """
+        test iterator over the entity set
+        """
+        from_iter = {d['id']: d for d in Blob()}
+        assert_equal(len(from_iter), len(Blob()))
+        assert_equal(from_iter[1]['blob'], 'character string')
+